feat: Add function to get DisProt data
DisProt has an api to get data for a protein based on its id.
The `show_ambiguous` and `show_obsolete` options are set to true.
These values give the most complete data.
This choice may later be revised to improve data quality.
The amino acid sequence is the desired predictor data, `X`.
The type label for each amino acid is the desired target data, `Y`.
A helper function is constructs that sequence from the raw return data.
Both sequences are returned as numpy unicode arrays.
Uncharacterized amino acids are given the label `?`.