feat: Add method to return data sets
This method return four lists of X and Y data for training and testing.
An 80/20 train/test split is used.
The load_data_sets method attempts to import that data from a file.
If the file does not exist the api requests are used to gather the data.
The api methods are replaced with mocks for testing get_data_sets.