bug: Change data type to match assert type
A previous commit changed the return type of load_data_sets to a list.
The assertion for the relevant test was changed to assertListEqual.
The variable for the expected value was not converted to a list.
This caused the test to fail.
The type of expected_data was changed to a list to fix the error.