bug: Change Y flatten to concatenate
In the ModalValueClassifier's fit method Y is a collection of sequences.
Fit calls the stats.mode function on the aggregate of Y values.
The sequences are 1D arrays, but their lengths need not be consistent.
Therefore, Y can be a 1D array of 1D arrays rather than a 2D array.
In that case calling Y.flatten() fails at aggregating the values.
Instead np.concatenate is used.