feat: Add ignore values to modal fit for arrays
The fit method of modal value classifier let some values be ignored.
There are 2 ways to set the mode depending on whether Y is an array.
Ignoring values is already implemented when Y is not an array.
This change allows values to also be ignored when Y is an array.